requirements.txt
# Core AWS dependencies
aws-cdk-lib>=2.232.0
boto3>=1.42.0
botocore>=1.42.0

# Testing dependencies
pytest>=8.0.0
hypothesis>=6.140.0
pytest-cov>=4.0.0
moto>=4.2.0

# Utility dependencies
attrs>=25.0.0
python-dateutil>=2.9.0
